Introduce a temporary edge limit correction
Due to the logic used to define e_min and e_max, e_min often appeared larger than e_max, which broke the logic of initialising e. Temporarily introduced a fix that checks if this is the case, and if so, swaps their values. Eventually, the calculation of e_min and e_max should be corrected.