Update al_cont.m
Unsure why padding is needed when calculating size. Removing addresses a warning related to updating FaceVertexAlphaData when in sphere mode. Couldn't see an issue arising from removing padding from the size calculation. If padding should be included in this size calculation, then the method for updating FaceVertexAlphaData in updateuiinfo_selection needs to be corrected instead.